Adding forked code from FryProxy
This commit adds the initial code as forked from the FryProxy project.
It was forked as of commit 7f384823bc579008d483d1534e5b435a668e6efb,
and will be maintained separately from this point forward. If the
original author would be willing to make the changes I've made in
this fork (removing the dependency on log4net and removing the use
of code contracts), I would gladly remove this fork.