Modifying processing pipeline to be thread-safe
When receiving multiple processing requests, there is a chance they
may interfere with one another. Until we can properly make the entire
operation thread-safe, we will lock the processing pipeline for a single
HTTP request in a mutex. This means that the proxy can only completely
process a single request at a time, which will be a performance bottleneck.
A rearchitecture to make the entire pipeline thread-safe should be
implemented as soon as practical.