Updating the header value separator to CRLF
Since header values may contain commas as a legitimate part of the
value, multiple instances of a header in a response are now separated
by carriage return/line feed (CRLF) pairs instead of commas. This only
affects the display of values when multiple instances of a single
header are returned.

Bumping the version to 1.1.1 with this commit.