Fixing read and write of responses using chunked encoding
Previously, after reading the results of a chunk, the proxy tried to
read the length of the next chunk, by reading the next line of the body
stream. However, if there are no more chunks, this could cause the
reader to attempt reading past the end of the stream, causing an
exception. To mitigate this, after reading a chunk, the proxy now
advances the stream pointer two bytes, past the carriage-return/line-
feed pair (CRLF) that marks the end of a chunk. The proxy can then
properly determine if it's reached the end of the stream before
attempting to read another chunk.